fix(clerk-js): Split component contexts into separate files
