+++ .changeset/tender-hairs-scream.md
@@ -1,0 +1,5 @@
+---
+'@clerk/clerk-js': patch
+---
+
+Split component contexts into their own files. This change does not impact consumers.
