+++ packages/clerk-js/src/ui/contexts/utils.ts
@@ -1,6 +1,8 @@
+import { snakeToCamel } from '@clerk/shared/underscore';
 import type { Clerk } from '@clerk/types';
 
 import { clerkCoreErrorContextProviderNotFound, clerkCoreErrorNoClerkSingleton } from '../../core/errors';
+import { createDynamicParamParser } from '../../utils';
 
 export function assertClerkSingletonExists(clerk: Clerk | undefined): asserts clerk is Clerk {
   if (!clerk) {
@@ -13,3 +15,17 @@ export function assertContextExists(contextVal: unknown, providerName: string): 
     clerkCoreErrorContextProviderNotFound(providerName);
   }
 }
+
+export function getInitialValuesFromQueryParams(queryString: string, params: string[]) {
+  const props: Record<string, string> = {};
+  const searchParams = new URLSearchParams(queryString);
+  searchParams.forEach((value, key) => {
+    if (params.includes(key) && typeof value === 'string') {
+      props[snakeToCamel(key)] = value;
+    }
+  });
+
+  return props;
+}
+
+export const populateParamFromObject = createDynamicParamParser({ regex: /:(\w+)/ });
